Feat : 가게에 리뷰 추가하기 API

+++ src/main/java/umc/study/web/dto/ReviewRequestDTO.java
@@ -12,7 +12,7 @@ public class ReviewRequestDTO {
         @NotNull
         Long storeId;
         @NotNull
-        Long score;
+        Float score;
         @NotNull
         String text;
     }
